+++ src/helpers/uuid.ts
@@ -1,0 +1,6 @@
+/**
+ * Cheap solution to generate a unique ID
+ */
+export function generateUID(): string {
+  return `${performance.now()}-${Math.random().toString(36).slice(2)}`;
+}
